#1745 moving apolloconfig and directives
