+++ components/rmrk/Gallery/GalleryItem.vue
@@ -206,7 +206,7 @@ import { get, set } from 'idb-keyval'
 import { MediaType } from '../types'
 import axios from 'axios'
 import { exist } from './Search/exist'
-import Orientation from '@/directives/DeviceOrientation'
+import Orientation from '@/utils/directives/DeviceOrientation'
 import PrefixMixin from '~/utils/mixins/prefixMixin'
 
 @Component<GalleryItem>({
