+++ components/unique/Gallery/Item/GalleryItem.vue
@@ -219,7 +219,7 @@ import { fetchNFTMetadata } from '@/components/rmrk/utils'
 import { get, set } from 'idb-keyval'
 import { MediaType } from '@/components/rmrk/types'
 import axios from 'axios'
-import Orientation from '@/directives/DeviceOrientation'
+import Orientation from '@/utils/directives/DeviceOrientation'
 import SubscribeMixin from '@/utils/mixins/subscribeMixin'
 import Connector from '@vue-polkadot/vue-api'
 import { Option } from '@polkadot/types'
